Update call to AsyncTask on swipe to recipes tab
Grab a list of checked ingredients from the parent activity and update
the url that will be used to make the api request. Also re-call
the asynctask's execute on swipe event. This is working normally, but
when user unchecks one ingredient, the other ingredient is automatically
"checked"; i.e. its isSelected() returns true although it is not checked
in the list view.