Sync checkbox and textview clicks
There was a bug where we had the checkbox onclick listener in our
ingredients adapter while our textview onclick listener was in our
ingredient fragment. Both appeared to sync up but they weren't in
the background.
Move the listener for the textview into the ingredients adapter.
Handle all click events and send the list of selected ingredients
over to the ingredients fragment.