Add placeholder as recipe images download
Use placeholder image as recipe images are downloaded.

Handle the error related to swiping from the ingredients fragment
quickly to the recipes fragment: "The content of the adapter has changed
but ListView did not receive a notification. Make sure the content of
your adapter is not modified from a background thread, but only from the
UI thread."

Add vodka to the search suggestions per Oriah's request. He's a
recovering alcoholic.